backend/app/schedules/serializers.py
from rest_framework import serializers
from .models import Schedule

class ScheduleSerializer(serializers.ModelSerializer):
    class Meta:
        model = Schedule
        fields = ['id', 'activity', 'day_of_week', 'start_time', 'end_time', 'details']
