backend/app/requirements.txt
Django==4.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
psycopg2==2.9.7
gunicorn==20.1.0
django-cors-headers==3.13.0
whitenoise